Add first test

+++ source/main.cpp
@@ -2,5 +2,5 @@
 #include <iostream>
 
 int main(int argc, char const* argv[]) {
-    return (int)scftolus::Main(argc, argv);
+    return scftolus::Main(argc, argv);
 }
